Use Application keyWindow when displaying from a BarButtonItem
The the ViewController happens to be a TableViewController then when
presenting from a BarButtonItem the topViewController was being
selected at the containerWindow. The problem with this in the case of a
TableViewController is that the PopTip view would scroll with the
table.

Suggested solution here is to use the Applications keyWindow in this
case.